pyproject.toml
[tool.poetry]
name = "integrationagent"
version = "0.1.0"
description = ""
authors = ["alanalanlu <alanlu1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.2.0"
langchain-core = "^0.3.1"
langgraph = "^0.2.22"
langsmith = "^0.1.122"
python-dotenv = "^1.0.1"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
